move: all analysing methods from main.py into htcanalyser.py accessible via the HTCAnalyser class.
